refactors ui to use ee stale endpoint
