+++ frontend/src/metabase-types/api/collection.ts
@@ -128,8 +128,6 @@ export interface CollectionItem {
   setCollectionPreview?: (isEnabled: boolean) => void;
 }
 
-export type StaleCollectionItem = CollectionItem & { last_used_at: string };
-
 export interface CollectionListQuery {
   archived?: boolean;
   "exclude-other-user-collections"?: boolean;
@@ -156,17 +154,6 @@ export type ListCollectionItemsRequest = {
 export type ListCollectionItemsResponse = {
   data: CollectionItem[];
   models: CollectionItemModel[] | null;
-} & PaginationResponse;
-
-export type ListStaleCollectionItemsRequest = {
-  id: CollectionId;
-  before_date?: string;
-  is_recursive?: boolean;
-} & PaginationRequest &
-  Partial<SortingOptions>;
-
-export type ListStaleCollectionItemsResponse = {
-  data: StaleCollectionItem[];
 } & PaginationResponse;
 
 export interface UpdateCollectionRequest {
